reviewid:10034 内核点播的版本号修改为uint16
git-svn-id: http://svn01/repositories/r01/p2p/trunk@2946 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ p2sp/download/DownloadDriver.h
@@ -32,7 +32,7 @@ namespace p2sp
     {
         Guid                  gPeerID;                              // B: 华哥ID
         Guid                  gResourceID;                          // C: ResourceID
-        boost::uint8_t        aPeerVersion[4];                      // D: 内核版本：major, minor, micro, extra
+        boost::uint16_t       aPeerVersion[4];                      // D: 内核版本：major, minor, micro, extra
         char                  szVideoName[512];                     // E: 视频名称/WCHAR
         char                  szOriginalUrl[1000];                  // F: Url
         char                  szOriginalReferUrl[1000];             // G: Refer Url
